Refactor server.ts file
